Refactor FeedbackListFragment sync logic to FeedbackListViewModel
Moved sync logic (start sync, check server, manage progress) from Fragment to ViewModel.
- Injected `SyncManager`, `ServerUrlMapper`, and `SharedPreferences` into `FeedbackListViewModel`.
- Implemented `startFeedbackSync` in ViewModel using Coroutines.
- Exposed `syncStatus` as `StateFlow` by observing `SyncManager.syncStatus`.
- Updated `FeedbackListFragment` to observe `syncStatus` and update UI.
- Removed legacy sync code from Fragment.
- Fixed an issue where `ImprovedSyncManager` would not update `SyncManager` status by wrapping the listener.
- Fixed `isShowing` compilation error in Fragment.
